Increasing code coverage
Also, fixing up MPI-based unit tests, and adding them back into the
code coverage reporting on GitHub.
