+++ .coveragerc
@@ -20,6 +20,7 @@ exclude_lines =
     raise AssertionError
     raise KeyboardInterrupt
     raise NotImplementedError
+    except ImportError
 
     # Don't complain if non-runnable code isn't run:
     if __name__ == .__main__.:
