+++ .github/workflows/coverage.yaml
@@ -24,5 +24,5 @@ jobs:
         run: pip install tox
       - name: Run Tox
         # Run tox using the version of Python in `PATH`
-        run: tox -e cov,report
+        run: tox -e mpicov,report
