+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py37,lint,cov
+envlist = py37,lint,cov,mpicov
 # need pip at least with --prefer-binary
 requires =
 	pip >= 20.2
@@ -7,6 +7,7 @@ requires =
 [testenv]
 basepython = {env:PYTHON3_PATH:python3}
 deps=
+    pip >= 20.2
     -r{toxinidir}/requirements.txt
     -r{toxinidir}/requirements-testing.txt
 setenv =
@@ -25,10 +26,27 @@ commands =
 
 [testenv:cov]
 deps=
+    pip>=20.2
     -r{toxinidir}/requirements.txt
     -r{toxinidir}/requirements-testing.txt
 commands =
     pytest --cov-config=.coveragerc --cov=armi --ignore=armi/utils/tests/test_gridGui.py {posargs} armi
+
+# TODO: So far this will only work in BASH Linux
+[testenv:mpicov]
+deps=
+    pip >= 20.2
+    mpi4py
+setenv =
+    COVERAGE_FILE = .coverage.mpi
+whitelist_externals =
+    /usr/bin/mpiexec
+    /usr/bin/mpiexec.exe
+commands =
+    pip install -r requirements.txt
+    pip install -r requirements-testing.txt
+    pytest --cov-config=.coveragerc --cov=armi --cov-append --ignore=armi/utils/tests/test_gridGui.py {posargs} armi
+    mpiexec -n 2 pytest --cov-config=.coveragerc --cov=armi --cov-append {posargs} armi/tests/test_mpiFeatures.py
 
 [testenv:lint]
 ignore_errors = true
@@ -40,6 +58,7 @@ skip_install = true
 deps=
     coveralls
 commands =
+    coverage combine
     coverage report
     coverage html
     coveralls --service=github
